requirements.txt
gradio>=4.0.0
faiss-cpu>=1.7.0
numpy>=1.21.0
pickle-mixin>=1.0.2
google-generativeai>=0.3.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0 